all radiuses added, clean up code, result class model added

+++ src/main/java/com/example/nukebombsimulation/calculations/EnergyParser.java
@@ -3,7 +3,7 @@ package com.example.nukebombsimulation.calculations;
 public class EnergyParser {
     private static final int ENERGY = 1000000;
 
-    public int parse(int bombMass){
+    public double parse(double bombMass){
         return bombMass * ENERGY;
     }
 }
